queries to separate files
