+++ src/containers/Tag/TagList/TagList.tsx
@@ -18,28 +18,10 @@ import EditIcon from '@material-ui/icons/Edit';
 import { Tag } from '../../../store/Tag/types';
 import styles from './TagList.module.css';
 
-const GET_TAGS = gql`
-  {
-    tags {
-      id
-      label
-      description
-    }
-  }
-`;
+import { GET_TAGS } from '../../../graphql/queries/Tag';
+import { DELETE_TAG } from '../../../graphql/mutations/Tag';
 
-const DELETE_TAG = gql`
-  mutation deleteTag($id: ID!) {
-    deleteTag(id: $id) {
-      errors {
-        key
-        message
-      }
-    }
-  }
-`;
-
-export interface TagListProps { }
+export interface TagListProps {}
 
 export const TagList: React.SFC<TagListProps> = (props) => {
   const [newTag, setNewTag] = useState(false);
@@ -48,7 +30,7 @@ export const TagList: React.SFC<TagListProps> = (props) => {
 
   let deleteId: number = 0;
   const [deleteTag] = useMutation(DELETE_TAG, {
-    update(cache, { data: { deleteTag } }) {
+    update(cache) {
       const tags: any = cache.readQuery({ query: GET_TAGS });
       const tagsCopy = JSON.parse(JSON.stringify(tags));
       tagsCopy.tags = tags.tags.filter((val: any) => val.id !== deleteId);
